font-patcher: Add possibility to create single width icons without touching existing glyphs
[why]
This can help if you want monospaced icons but not-force the other
glyphs to be monospaced (which we do to make the whole font
monospace-detectable which was a major issue in the beginning, esp with
Windows).

[how]
Add option --single-width-glyphs that makes the added glyphs single
width (like --mono), but leaves preexisting glyphs untouched.

Fixes: #1772

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>