font-patcher: Add new option --cell
EXPERIMENTAL

[why]
Sometimes users want to tweak the cell sizing or the baseline to
baseline distance, or the middle point of the cell.

[how]
Add a new option to show the 'cell' box and to override the detected
one. It is not sufficient to adjust width and height, because that can
not define shifts of the cell up/down (left/right is mostly useless and
I believe the code does not work if the xmin is not zero).

The smaller icon-height is not used here (affecting only --mono) because
that seems to compilcated right now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>